fix(web): prevent tree page from showing blank screen
When loading for the first time the `/tree` page  shows blank space without even navbar and footer. This is due to dynamic loading of the page content.

Here I:
 - bubble the `Layout` component from tree page one level up, out of dynamically loaded chunk, so that navbar and footer are static and displayed at all times
 -  add `loading` parameter to `dynamic()` call, such that the loading spinner is displayed during page fetching

The page still freezes after fetching, while the tree is rendered. I believe this is due to the sheer amount of things being rendered in Auspice, but might also be perf problems on Nextclade side. I don't currently know a way to handle this or how to display loading indicator there, because React rendering is still synchronous and blocks main thread.
