+++ packages_rs/nextclade-web/src/pages/tree.tsx
@@ -1,5 +1,1 @@
-import dynamic from 'next/dynamic'
-
-const TreePage = dynamic(() => import('src/components/Tree/TreePage'), { ssr: false })
-
-export default TreePage
+export { TreePage as default } from 'src/components/Tree/TreePage'
